+++ lcs/agents/racs/Configuration.py
@@ -16,7 +16,9 @@ class Configuration:
                  theta_r=0.9,
                  epsilon=0.5,
                  u_max=100000,
-                 theta_exp=20,) -> None:
+                 theta_exp=20,
+                 theta_ga=100,
+                 theta_as=20) -> None:
 
         self.oktypes = (UBR,)
         self.encoder = RealValueEncoder(encoder_bits)
@@ -37,3 +39,5 @@ class Configuration:
         self.u_max = u_max
 
         self.theta_exp = theta_exp
+        self.theta_ga = theta_ga
+        self.theta_as = theta_as